New structure for making it easier to import the dream-wallet core modules
