+++ src/index.js
@@ -1,13 +1,15 @@
-import { walletReducer } from './redux/reducers'
-// import { walletSyncMiddleware } from './walletSyncMiddleware'
-import * as walletActions from './redux/actions'
-import * as Network from './network'
+import * as coreActions from './redux/actions'
+import * as coreMiddleware from './redux/middleware'
+import * as coreReducers from './redux/reducers'
 import * as coreSelectors from './redux/selectors'
+import * as Network from './network'
+import * as coreSagas from './redux/sagas.js'
 
 export {
-  walletReducer,
-  // walletSyncMiddleware,
-  walletActions,
-  Network,
-  coreSelectors
+  coreActions,
+  coreMiddleware,
+  coreReducers,
+  coreSelectors,
+  coreSagas,
+  Network
 }
