+++ src/redux/data/LatestBlock/actions.js
@@ -1,0 +1,4 @@
+export const LATEST_BLOCK_DATA_LOAD = '@v3.LATEST_BLOCK_DATA_LOAD'
+
+export const loadLatestBlockData = (data) =>
+  ({ type: LATEST_BLOCK_DATA_LOAD, payload: data })
