+++ src/redux/data/Wallet/selectors.js
@@ -1,0 +1,1 @@
+export const getBalance = bd => bd.wallet.final_balance
