+++ src/redux/middleware.js
@@ -1,0 +1,7 @@
+import walletSync from './wallet/middleware/walletSync.js'
+import socket from './wallet/middleware/socket.js'
+
+export {
+  socket,
+  walletSync
+}
