--- src/redux/middleware/index.js
@@ -1,3 +1,0 @@
-
-export { default as walletSyncMiddleware } from './walletSyncMiddleware'
-export { default as walletSocketMiddleware } from './walletSocketMiddleware'
