--- src/redux/reducers/index.js
@@ -1,2 +1,0 @@
-export { default as walletReducer } from './walletReducer'
-export { default as blockchainDataReducer } from './blockchainDataReducer'
