+++ src/redux/wallet/selectors.js
@@ -1,0 +1,7 @@
+import { compose } from 'ramda'
+import { Wallet, Wrapper } from '../../types'
+
+const ImtoJS = i => i.toJS()
+export const getWalletContext = compose(ImtoJS, Wallet.selectContext, Wrapper.selectWallet)
+export const getAddressContext = compose(ImtoJS, Wallet.selectAddrContext, Wrapper.selectWallet)
+export const getXpubsContext = compose(ImtoJS, Wallet.selectXpubsContext, Wrapper.selectWallet)
